refactor: 페이지 관련 컴포넌트를 pages 디렉터리 하위로 이동
